% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_vec_mat.R
\name{bcolumn_vector}
\alias{bcolumn_vector}
\title{Convert Column Vector To bmatrix in Rmarkdown}
\usage{
bcolumn_vector(pvec, ps_name = NULL, ps_equal_sign = "=", ps_env = NULL)
}
\arguments{
\item{pvec}{vector to be converted to Rmarkdown}

\item{ps_name}{variable name for the vector}

\item{ps_equal_sign}{equal sign to be used when a variable name is given}

\item{ps_env}{math environment in which matrix should be presented}
}
\value{
String containing bmatrix representation of given vector
}
\description{
The specified column vector is converted to a bmatrix in Rmarkdown
}
\details{
#
---- bcolumn_vector Wrapper --------------------------------------------- #
}
\examples{
\dontrun{
example_vec <- c(1:5)
bcolumn_vector(pvec = example_vec, ps_name = 'v', ps_env = '$$')
}

}
