% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_helper.R
\name{beamer_with_notes}
\alias{beamer_with_notes}
\title{Render beamer presentation with notes}
\usage{
beamer_with_notes(ps_input_path,
  ps_output_file = paste0(tools::file_path_sans_ext(basename(ps_input_path)),
  "_with_notes.pdf"), ps_in_header = "header_show_notes.tex")
}
\arguments{
\item{ps_input_path}{rmarkdown source file with note sections}

\item{ps_output_file}{pdf output file}

\item{ps_in_header}{latex header file}
}
\description{
Use the option to add notes to beamer presentations via an
included header file that tells latex to include the notes.
}
\examples{
\dontrun{
beamer_with_notes(ps_input_path = 'sl/w01/lbgfs2019_slides_w01_intro-lbg.Rmd')
}
}
