% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmatrix.R, R/rmd_vec_mat.R
\name{bmatrix}
\alias{bmatrix}
\title{Convert R-matrix to a bmatrix}
\usage{
bmatrix(pmat, ps_name = NULL, ps_equal_sign = "=", ps_env = NULL)

bmatrix(pmat, ps_name = NULL, ps_equal_sign = "=", ps_env = NULL)
}
\arguments{
\item{pmat}{matrix to be converted to Rmarkdown}

\item{ps_name}{variable name for the matrix}

\item{ps_equal_sign}{equal sign to be used when a variable name is given}

\item{ps_env}{math environment in which matrix should be presented}
}
\value{
String containing bmatrix representation of given matrix
}
\description{
According to https://bookdown.org/yihui/rmarkdown/markdown-syntax.html 
bmatrix is the environment that encloses the matrix into brackets

The specified matrix is converted to a bmatrix in Rmarkdown
}
\details{
#
---- bmatrix Wrapper ---------------------------------------------------- #
}
\examples{
\dontrun{
eye3 <- diag(1, nrow = 3)
bmatrix(pmat = eye3, ps_name = 'I', ps_env = '$$')
}

}
