% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{clean_rmd_dir}
\alias{clean_rmd_dir}
\title{Clean RMarkdown Directory}
\usage{
clean_rmd_dir(ps_rmd_dir)
}
\arguments{
\item{ps_rmd_dir}{RMarkdown source directory}
}
\description{
RMarkdown docments are often placed in separate directories to avoid cluttering
a given directory with output files of different source files. Output files
are almost never versioned, hence it is useful to have a function to remove
all RMarkdown output files. This function helps to clean up directories from
RMarkdown output files.
}
\examples{
\dontrun{
clean_rmd_dir(ps_rmd_dir = 'rmd')
}

}
