% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{convert_odg}
\alias{convert_odg}
\title{Converter Function from Odg to Other Formats}
\usage{
convert_odg(ps_odg_path, ps_out_format)
}
\arguments{
\item{ps_odg_path}{path to odg file to be converted}

\item{ps_out_format}{output format into which ps_odg_path should be converted to}
}
\description{
The conversion is done using the tool returned by \code{get_odg_prog_path()}.
The current version of this function is not vectorized, hence the arguments
can only be single odg-files in ps_odg_path and single formats in ps_out_format.
The conversion is only done, if the result file does not exist or if the
result file has an older modification time as found by \code{file.mtime()}.
}
