% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-root.R
\name{create_ghroot_project}
\alias{create_ghroot_project}
\title{Create a gh-root project}
\usage{
create_ghroot_project(ps_project, ps_path = ".", pb_verbose = FALSE)
}
\arguments{
\item{ps_project}{projectname}

\item{ps_path}{path under which project is created}
}
\description{
GH-root projects that have a master branch and a gh-pages
branch in different subdirectories under a common root
which is called the gh-root.
}
