% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_helper.R
\name{create_header_hide_notes}
\alias{create_header_hide_notes}
\title{Create latex header file to hide notes}
\usage{
create_header_hide_notes(ps_input_dir,
  ps_in_header = "header_hide_notes.tex")
}
\arguments{
\item{ps_input_dir}{rmarkdown input directory}

\item{ps_in_header}{name of the latex header file}
}
\description{
Create latex header file to hide notes
}
\examples{
\dontrun{
create_header_hide_notes(ps_input_dir = 'sl/w01')
}
}
