% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{get_latest_change}
\alias{get_latest_change}
\title{Write data-time and user that has done latest changes to a document}
\usage{
get_latest_change(ps_msg = "Latest Changes")
}
\arguments{
\item{ps_msg}{special message to be used in front of time and user info, defaults to Latest Changes}
}
\value{
st_result containing time and user-part of system info
}
\description{
The information about who has done the latest changes to a document
is usually added at the end of a website. This allows the reader
of a website to check whether the site was recently updated. For
the author, it is an easy way to check whether the most recent
version has been uploaded to the server.
}
\details{
The information that is returned by this function consists of the
time returned by \code{Sys.time()} and the username that is returned
by \code{Sys.info()}. The intended way to use this function is to
\code{cat()} the result of the function in the last code-junk of
a document. See examples for more details.
}
\examples{
\dontrun{
```{r}
cat(rmddochelper::get_latest_change(), "\\n")
```
}

}
