% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendeley_export_r6class.R
\name{get_met}
\alias{get_met}
\title{Wrapper to create a MendeleyExportToolR6 Object}
\usage{
get_met()
}
\value{
result_met MendelyExportToolR6 object
}
\description{
This wrapper creates a MendelyExportToolR6 object and returns a
reference to that object as a result. All parameters for the
creation of the MendelyExportToolR6 are used at their default
values.
}
