% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{get_odg_prog_path}
\alias{get_odg_prog_path}
\title{Return Path To Program Used To Create ODG-graphics}
\usage{
get_odg_prog_path()
}
\description{
The path to the program used to create odg-graphics depends
on the OS. We use the function get_os() to determine on which
OS, we are running and based on that, we are returning a
fixed string corresponding to the path to the program for
odg-graphics. In most cases,
LibreOffice draw is used to create the odg-graphics files.
}
\examples{
\dontrun{
get_odg_prog_path()
# returns /Applications/LibreOffice.app/Contents/MacOS/soffice on OSX
}
}
