% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{get_os}
\alias{get_os}
\title{Get name of OS we are running on}
\usage{
get_os()
}
\value{
name of operating system which is either
  "windows", "osx" or "linux".
}
\description{
This function returns a simple string depending whether
the current instance of R runs on windows, mac osx or
linux.
}
\details{
The function was proposed on the blog post at
\url{http://conjugateprior.org/2015/06/identifying-the-os-from-r/}
found on R-bloggers.

The case of windows was not included in the original
version from the blog post. It was included here
at the beginning to filter them out at the beginning
based on the value of \code{.Platform$OS.type}
}
