% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{get_pat_pos}
\alias{get_pat_pos}
\title{Get position where search pattern occurs in a file}
\usage{
get_pat_pos(ps_file, ps_pattern)
}
\arguments{
\item{ps_file}{name of the file to search for pattern}

\item{ps_pattern}{search pattern for which we search in the file}
}
\value{
vec_pos_found vector of positions where pattern occurs,
                      NA if pattern was not found
}
\description{
Given a search pattern and given a file that is specified
by its complete path, the positions where the search
pattern occurs is returned. Positions correspond to
vector indices when the content of the file is read
using the function \code{readLines}. These vector indices
are equivalent to line numbers of the text. In case
when the search pattern is not found, NA is returned.
}
