% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{get_rmd_src_with_pos}
\alias{get_rmd_src_with_pos}
\title{Return rmd source file with positions where search pattern occurs}
\usage{
get_rmd_src_with_pos(ps_path, ps_cwd)
}
\arguments{
\item{ps_path}{name of and path to odg-graphics file}

\item{ps_cwd}{current working directory}
}
\value{
l_rmd_src_result list with name of rmd source file and positions where search pattern occurs
}
\description{
Given a search pattern in ps_path and given the path of the current
working directory, all files with extension .Rmd are searched whether
the string in ps_path occurs in the rmd-source file. If the pattern
is found in exactly one Rmd-source file, the name of the Rmd-source
file and the positions where the pattern was found is returned. If
the pattern is not found or the pattern occurs in more than one
Rmd-source file, NULL is returned.
}
