% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{get_this_rmd_file}
\alias{get_this_rmd_file}
\title{Get the name of the current rmd file}
\usage{
get_this_rmd_file()
}
\value{
name of current rmd file
}
\description{
Depending on whether this function is called from
inside of RStudio or from outside, the rstudioapi
is used or rprojroot is used to determine the
name of the current rmdfile
}
