% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{hook_convert_odg}
\alias{hook_convert_odg}
\title{Odg Graphics File Conversion Hook}
\usage{
hook_convert_odg(before, options, envir)
}
\arguments{
\item{before}{flag whether hook is run before or after chunk}

\item{options}{list of options passed from chunk to hook}

\item{envir}{environment}
}
\description{
Graphics objects in odg-format are converted to png
or pdf format by this knitr-hook-function. We assume
that this hook-function is included in the chunk where
the graphics object is included using a call to
\code{knitr::include_graphics()}, hence the conversion
of the odg-file must be run before the code in the chunk.
}
\examples{
\dontrun{
# registration of hook function
knitr::knit_hooks$set(hook_convert_odg = rmddochelper::hook_convert_odg)
...
}

}
