% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{insert_include_command}
\alias{insert_include_command}
\title{Insert Graphics Inclusion Command Into Rmd-Source}
\usage{
insert_include_command(ps_path, ps_rmd_src = NULL, ps_cwd = get_wd())
}
\arguments{
\item{ps_path}{path to the odg graphics file}

\item{ps_rmd_src}{name of and path to the rmd-source file}

\item{ps_cwd}{current working directory of the rmd-source file}
}
\description{
When creating a new graphics object, this function tries to automatically
generate the associated command to include the generated graphics object
into the respective Rmd-source document. This process needs as input
the current working directory where of the Rmd-source document and the
name of the file in which the graphics object is stored.
}
