% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helper.R
\name{matDiag}
\alias{matDiag}
\title{Diagonal matrix}
\usage{
matDiag(psBaseElement, pnNrRow, pnNrCol)
}
\arguments{
\item{psBaseElement}{constant prefix of each matrix element}

\item{pnNrRow}{number of rows}

\item{pnNrCol}{number of columns}
}
\description{
Diagonal matrix
}
