% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_r6class.R
\docType{class}
\name{R6ClassCount}
\alias{R6ClassCount}
\title{R6 Class representing object counts in documents}
\usage{
R6ClassCount$new()
}
\description{
Simple counter of objects inside a document. Objects that
associated to such a counter can be sections, tables,
diagrams and others. Specialised counter classes for the
different document objects can be derived from this
more generic counter class. Specific document objects
that are associated to a counter are generated using
the prefix field in this class.
}
\section{Public methods}{

\describe{
  \item{\code{new()}}{Instantiation of a new R6ClassCount object}
  \item{\code{out(ps_prefix, ps_suffix, pn_count)}}{Write count value with given prefix and suffix}
  \item{\code{increment_count()}}{Increment value of count by 1}
  \item{\code{set_prefix(ps_prefix)}}{Setter for field prefix}
  \item{\code{get_prefix()}}{Getter for field prefix}
  \item{\code{set_suffix(ps_suffix)}}{Setter for field suffix}
  \item{\code{get_suffix()}}{Getter for field suffix}
  \item{\code{set_count(pn_count)}}{Setter for field count}
  \item{\code{get_count()}}{Getter for field count}
}
}

\examples{
cnt <- R6ClassCount$new()
cat(cnt$out(ps_prefix = "## Task", ps_suffix="Variance components", pn_count = 1))

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{Text shown before counter, specifies the document object type}

\item{\code{suffix}}{Test shown after the counter, corresponds to a description}

\item{\code{count}}{current value of the counter}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{prefix}}{Text shown before counter, specifies the document object type}

\item{\code{suffix}}{Test shown after the counter, corresponds to a description}

\item{\code{count}}{current value of the counter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R6classCount-new}{\code{R6ClassCount$new()}}
\item \href{#method-R6classCount-out}{\code{R6ClassCount$out()}}
\item \href{#method-R6classCount-increment_count}{\code{R6ClassCount$increment_count()}}
\item \href{#method-R6classCount-set_prefix}{\code{R6ClassCount$set_prefix()}}
\item \href{#method-R6classCount-get_prefix}{\code{R6ClassCount$get_prefix()}}
\item \href{#method-R6classCount-set_suffix}{\code{R6ClassCount$set_suffix()}}
\item \href{#method-R6classCount-get_suffix}{\code{R6ClassCount$get_suffix()}}
\item \href{#method-R6classCount-set_count}{\code{R6ClassCount$set_count()}}
\item \href{#method-R6classCount-get_count}{\code{R6ClassCount$get_count()}}
\item \href{#method-R6classCount-clone}{\code{R6ClassCount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-new"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-out"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-out}{}}}
\subsection{Method \code{out()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$out(ps_prefix = NULL, ps_suffix = NULL, pn_count = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-increment_count"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-increment_count}{}}}
\subsection{Method \code{increment_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$increment_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-set_prefix"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-set_prefix}{}}}
\subsection{Method \code{set_prefix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$set_prefix(ps_prefix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-get_prefix"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-get_prefix}{}}}
\subsection{Method \code{get_prefix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$get_prefix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-set_suffix"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-set_suffix}{}}}
\subsection{Method \code{set_suffix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$set_suffix(ps_suffix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-get_suffix"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-get_suffix}{}}}
\subsection{Method \code{get_suffix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$get_suffix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-set_count"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-set_count}{}}}
\subsection{Method \code{set_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$set_count(pn_count)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-get_count"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-get_count}{}}}
\subsection{Method \code{get_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$get_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R6classCount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R6classCount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6ClassCount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
