% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_vec_mat.R
\name{rmd_vec_mat}
\alias{rmd_vec_mat}
\title{Generic Converter of Matrices To Rmarkdown}
\usage{
rmd_vec_mat(
  pmat,
  ps_name = NULL,
  ps_equal_sign = "=",
  ps_env = NULL,
  ps_out_type = "array",
  ps_out_format = "html"
)
}
\arguments{
\item{pmat}{matrix to be converted to Rmarkdown}

\item{ps_name}{variable name for the matrix}

\item{ps_equal_sign}{equal sign to be used when a variable name is given}

\item{ps_env}{math environment in which matrix should be presented}

\item{ps_out_type}{one of the output types available in Rmarkdown}

\item{ps_out_format}{output format produced}
}
\value{
sresult String containing Rmarkdown representation of given vector or matrix
}
\description{
The generic conversion takes a matrix and converts it into one
of the constructs recognized by Rmarkdown as described on
https://bookdown.org/yihui/rmarkdown/markdown-syntax.html.
}
\details{
---
title: Vectors and Matrices in Rmarkdown
date:  2020-05-25
---


This function is not called directly, but is used by different wrappers
for the different output types.
}
