% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helper.R
\name{sConvertMatrixToLaTexArray}
\alias{sConvertMatrixToLaTexArray}
\title{Convert matrix to LaTeX array}
\usage{
sConvertMatrixToLaTexArray(
  pmatAMatrix,
  pnOutStartLine = 5,
  pnEndIgnoreLines = 1,
  pnDigits = 2,
  pnAlign = NULL
)
}
\arguments{
\item{pmatAMatrix}{Matrix to be represented in tex format}

\item{pnOutStartLine}{line index where output should start, default = 5}

\item{pnEndIgnoreLines}{number of lines to be ignored at the end of the output, default = 1}

\item{pnDigits}{specify the number of digits to be used}

\item{pnAlign}{character vector denoting column alignment}
}
\value{
string containing tex representation of matrix
}
\description{
The matrix specified by pmatAMatrix is converted to a LaTeX array
using function \code{xtable::xtable}. The function \code{xtable::xtable}
produces a tabular LaTeX-object which is converted to a LaTeX-array using
simple string replacement of the LaTeX environment specifiers. Some
lines at the beginning and at the end are ignored. The number
of lines that are ignored can be specified using the parameters
pnOutStartLine and pnEndIgnoreLines.
}
\details{
Because in R vectors and matrices are different objects, when we want
to use this function for a vector, it has to be converted to a matrix
first.
}
