% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helper.R
\name{sConvertRowVectorToLaTexArray}
\alias{sConvertRowVectorToLaTexArray}
\title{Convert row vector to LaTeX array}
\usage{
sConvertRowVectorToLaTexArray(pvec_avector, pnDigits = 2)
}
\arguments{
\item{pvec_avector}{Vector to be converted}

\item{pnDigits}{specify the number of digits to be used}
}
\value{
row-vector converted to LaTeX-array
}
\description{
We are given a vector and it should be printed as a row vector
}
