% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helper.R
\name{sConvertVectorToLaTexArray}
\alias{sConvertVectorToLaTexArray}
\title{Convert vector to LaTeX array}
\usage{
sConvertVectorToLaTexArray(pvec_avector, pnDigits = 2)
}
\arguments{
\item{pvec_avector}{Vector to be converted}

\item{pnDigits}{specify the number of digits to be used}
}
\value{
vector converted to LaTeX-array
}
\description{
The specified vector is converted to a LaTeX array. We first
put the vector into a one column matrix and use sConvertMatrixToLaTexArray
on this one-column matrix
}
