% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odg_graphics.R
\name{show_knit_hook_call}
\alias{show_knit_hook_call}
\title{Show Knit Hook Command}
\usage{
show_knit_hook_call()
}
\description{
In every document that contains a graphic or diagram, we have to add the
odg-conversion function as knit-hook. This statement is not easy to remember
and therefore, we want a short cut with a simple function
}
\examples{
\dontrun{
show_knit_hook_call()
}
}
