% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_sh.R
\name{spin_sh}
\alias{spin_sh}
\title{Spin a bash script (goat's hair) into Rmd (wool)}
\usage{
spin_sh(hair, knit = TRUE, report = TRUE, text = NULL,
  envir = parent.frame(), doc = "^#+'[ ]?", comment = c("^[# ]*/[*]",
  "^.*[*]/ *$"), precious = !knit && is.null(text))
}
\arguments{
\item{hair}{path to the bash scirpt}

\item{comment}{pair of pattern defining a comment}

\item{precious}{should output source document be kept}
}
\description{
This function is derived from \code{knitr::spin()} for R-scripts and
is adapted to do a similar conversion of bash scripts into Rmd files.
This function (\code{spin_sh}) does not handle inline bash statements
which are bash commands inside a text. The only format that is
considered is Rmd.
}
