% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_sh.R
\name{split_lines}
\alias{split_lines}
\title{Split a string at newline into vector}
\usage{
split_lines(x)
}
\arguments{
\item{x}{original input string}
}
\value{
vector of lines
}
\description{
Split a string at newline into vector
}
