% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmatrix.R
\docType{data}
\name{tbl_breed_kb}
\alias{tbl_breed_kb}
\title{---
title: Experiments With Matrices
date:  2020-05-24
---}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
tbl_breed_kb
}
\description{
## Data
}
\keyword{datasets}
