% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_to_dir.R
\name{use_vignette_in_dir}
\alias{use_vignette_in_dir}
\title{Create Vignette Document in Sub-directory}
\usage{
use_vignette_in_dir(
  ps_name,
  ps_title = fs::path_ext_remove(ps_name),
  pb_open = FALSE
)
}
\arguments{
\item{ps_name}{Base of vignette file name}

\item{ps_title}{Title of vignette}

\item{pb_open}{Directly open the vignette}
}
\description{
A vignette is created in a separate subdirectory. The subdirectory by default is
the same as the file-name of the vignette. If the vignette name contains a
directory path, then it is saved under that directory.
}
\examples{
\dontrun{
use_vignette_in_dir("example-vignette", "Example Vignette")
}
}
