% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_helper.R
\name{vecGetVecElem}
\alias{vecGetVecElem}
\title{Vector with base elements and index}
\usage{
vecGetVecElem(psBaseElement, pnVecLen, psResult = "raw")
}
\arguments{
\item{psBaseElement}{suffix of the vector elements}

\item{pnVecLen}{number of elements in the vector}

\item{psResult}{format of result (default = "raw")}
}
\value{
Vector with string elements that each have psBaseElement as
        prefix and the index as suffix, separated by psSepChar
}
\description{
The result is a vector of elements that have psBaseElement
as prefix and that have the index of the corresponding
vector element as suffix. The result may be in raw format
or in LaTeX-Math format. The latter is produced when
argument psResult is set to "latex".
}
