% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_to_dir.R
\name{vig_to_dir}
\alias{vig_to_dir}
\title{Vignette Documents to Directories}
\usage{
vig_to_dir(
  pvec_rmd_doc = list.files(file.path(here::here(), "vignettes"), pattern = ".Rmd$"),
  ps_trg_dir = file.path(here::here(), "vignettes")
)
}
\arguments{
\item{pvec_rmd_doc}{Rmd document to be moved. By default all Rmd files in directory 'vignettes' are moved}

\item{ps_trg_dir}{alternative root target directory, by default: file.path(here::here(), "vignettes")}
}
\description{
Vignettes are documentation of packages in a paper-like format. The vignettes
can be created using the function \code{usethis::use_vignettes()}. This creates
a new RMarkdown (Rmd) document in the subdirectory 'vignettes'. So far it is not possible
to group different documents in separate directories. This function creates for
each Rmd document a separate directory and moves the Rmd document into that subdirectory.
}
\details{
The creation of the new subdirectory and the move of the created Rmd document
done using functions from package 'fs'.
}
\examples{
\dontrun{
usethis::use_vignettes("example-vignette")
vig_to_dir("vignettes/example-vignette.Rmd")
}
}
